requirements.txt
openai
jsonschema
pytest
google.generativeai
requests
beautifulsoup4
pandas
python-dotenv==1.0.1

